Move deconz_device to models
