+++ pydeconz/models/deconz_device.py
